Finished - Into Testing/Debugging Phase.
Works by taking in a zero-based phrasing system
- adding 1 to the read in number to make the system non-zero based (To allow 0 to be an escape char)
- Then feeds the number to the parsing method
- Figures out if the new number can fit into an existing byte, overflows from an existing byte or fits perfectly into an existing byte.
-Prints out
-Repeat until all numbers are read
-Print out any remaining.
The unpacker should make sure to minus the values by 1 before writing out. The unpacker should keep track of the max bits per number (can use the updateMaxBitsNum function to keep track)
The unpacker should read in 1 bit at a time to parse based on the correct numbers. The unpacker should make sure at end of file that 1. If it is a 0, ignore it, 2. If it doesn't fully count to the maxbitnum, ignore - that is end of transmission, 3. If there are no further bytes being printed, then end of transmission